Create README.md
Comparison of tissue-specific expression Visualizing differential expression with bar graphs and heat maps.
ferthermore can be interpretation of their biological roles through databases such as Gene Ontology (GO) and
UniProt. The analysis focused on energy metabolism-related mitochondrial genes that are important for both tissues.

before that to access this you have to download both file and import to jupyternotebook and run the ".ipynb" file 

feel free to share any false to me.